requirements.txt
stripe
pandas
python-dotenv
tqdm
openpyxl
phonenumbers